REFACTOR ALL THE THINGS
To elaborate: Rename all the page/tab group/box models (and their
associated Backbone things, routes, db tables, etc.) to be
templates. Things should be working (the tests pass, and the site
seems to be functional), but bugs may crop up since the test suite is
pretty minimal.

Next thing to do is create the filled-in version of all these models,
and work out the logic to make notebooks properly sharable.